Update tests to use OpenScmDataFrame
