+++ openscm/core/time.py
@@ -91,163 +91,6 @@ class InterpolationType(Enum):
         return interpolation_type
 
 
-def _float_year_to_datetime(inp: float) -> np.datetime64:
-    year = int(inp)
-    fractional_part = inp - year
-    return np.datetime64(  # pylint: disable=too-many-function-args
-        year - 1970, "Y"
-    ) + np.timedelta64(  # pylint: disable=too-many-function-args
-        int(
-            (
-                datetime.datetime(year + 1, 1, 1) - datetime.datetime(year, 1, 1)
-            ).total_seconds()
-            * fractional_part
-        ),
-        "s",
-    )
-
-
-_ufunc_float_year_to_datetime = np.frompyfunc(_float_year_to_datetime, 1, 1)
-_ufunc_str_to_datetime = np.frompyfunc(parser.parse, 1, 1)
-
-
-def _parse_datetime(inp: np.ndarray) -> np.ndarray:
-    try:
-        return _ufunc_float_year_to_datetime(inp.astype(float))
-    except (TypeError, ValueError):
-        return _ufunc_str_to_datetime(inp)
-
-
-def _format_datetime(dts: np.ndarray) -> np.ndarray:
-    """
-    Convert an array to an array of :class:`np.datetime64`.
-
-    Parameters
-    ----------
-    dts
-        Input to attempt to convert
-
-    Returns
-    -------
-    :class:`np.ndarray` of :class:`np.datetime64`
-        Converted array
-
-    Raises
-    ------
-    ValueError
-        If one of the values in :obj:`dts` cannot be converted to :class:`np.datetime64`
-    """
-    if len(dts) <= 0:  # pylint: disable=len-as-condition
-        return np.array([], dtype="datetime64[s]")
-
-    dtype = dts.dtype.type
-    if dts.dtype.kind == "O":
-        dtype = np.dtype(type(dts[0])).type
-    if issubclass(dtype, np.datetime64):
-        return np.asarray(dts, dtype="datetime64[s]")
-    if issubclass(dtype, np.floating):
-        return _ufunc_float_year_to_datetime(dts).astype("datetime64[s]")
-    if issubclass(dtype, np.integer):
-        return (np.asarray(dts) - 1970).astype("datetime64[Y]").astype("datetime64[s]")
-    if issubclass(dtype, str):
-        return _parse_datetime(dts).astype("datetime64[s]")
-    return np.asarray(dts, dtype="datetime64[s]")
-
-
-class TimePoints:  # TODO: track type of timeseries
-    """
-    Handles time points by wrapping :class:`np.ndarray` of :class:`np.datetime64`..
-    """
-
-    _values: np.ndarray
-    """Actual time points array"""
-
-    def __init__(self, values: Sequence):
-        """
-        Initialize.
-
-        Parameters
-        ----------
-        values
-            Time points array to handle
-        """
-        self._values = _format_datetime(np.asarray(values))
-
-    @property
-    def values(self) -> np.ndarray:
-        """
-        Time points
-        """
-        return self._values
-
-    def to_index(self) -> pd.Index:
-        """
-        Get time points as :class:`pd.Index`.
-
-        Returns
-        -------
-        :class:`pd.Index`
-            :class:`pd.Index` of :class:`np.dtype` :class:`object` with name ``"time"``
-            made from the time points represented as :class:`datetime.datetime`.
-        """
-        return pd.Index(self._values.astype(object), dtype=object, name="time")
-
-    def years(self) -> np.ndarray:
-        """
-        Get year of each time point.
-
-        Returns
-        -------
-        :obj:`np.array` of :obj:`int`
-            Year of each time point
-        """
-        return np.vectorize(getattr)(self._values.astype(object), "year")
-
-    def months(self) -> np.ndarray:
-        """
-        Get month of each time point.
-
-        Returns
-        -------
-        :obj:`np.array` of :obj:`int`
-            Month of each time point
-        """
-        return np.vectorize(getattr)(self._values.astype(object), "month")
-
-    def days(self) -> np.ndarray:
-        """
-        Get day of each time point.
-
-        Returns
-        -------
-        :obj:`np.array` of :obj:`int`
-            Day of each time point
-        """
-        return np.vectorize(getattr)(self._values.astype(object), "day")
-
-    def hours(self) -> np.ndarray:
-        """
-        Get hour of each time point.
-
-        Returns
-        -------
-        :obj:`np.array` of :obj:`int`
-            Hour of each time point
-        """
-        return np.vectorize(getattr)(self._values.astype(object), "hour")
-
-    def weekdays(self) -> np.ndarray:
-        """
-        Get weekday of each time point.
-
-        Returns
-        -------
-        :obj:`np.array` of :obj:`int`
-            Day of the week of each time point
-        """
-        return np.vectorize(datetime.datetime.weekday)(self._values.astype(object))
-
-
 def create_time_points(  # TODO: replace by simpler function
     start_time: np.datetime64,
     period_length: np.timedelta64,
