+++ openscm/parameter_type.py
@@ -1,0 +1,9 @@
+"""
+Functionality for guessing the parameter types from variable names and unit.
+"""
+import re
+from typing import Optional
+
+from openscm.core.parameters import ParameterType
+
+
