+++ tests/conftest.py
@@ -15,7 +15,7 @@ import pytest
 from openscm.core import time
 from openscm.core.parameters import ParameterType
 from openscm.core.parameterset import ParameterSet
-from openscm.scmdataframe import ScmDataFrame
+from openscm.scmdataframe import OpenScmDataFrame
 
 try:
     from pyam import IamDataFrame
@@ -164,7 +164,7 @@ def test_scm_datetime_df():
         inplace=True,
     )
 
-    yield ScmDataFrame(tdf)
+    yield OpenScmDataFrame(tdf)
 
 
 @pytest.fixture(scope="function")
@@ -212,12 +212,12 @@ def test_iam_df():
 def test_scm_df(request):
     if IamDataFrame is None:
         pytest.skip("pyam is not installed")
-    yield ScmDataFrame(**request.param)
+    yield OpenScmDataFrame(**request.param)
 
 
 @pytest.fixture(scope="function")
 def test_processing_scm_df():
-    yield ScmDataFrame(
+    yield OpenScmDataFrame(
         data=np.array([[1, 6.0, 7], [0.5, 3, 2], [2, 7, 0], [-1, -2, 3]]).T,
         columns={
             "model": ["a_iam"],
@@ -254,7 +254,7 @@ append_scm_df_pairs_times = [
     datetime(2010, 1, 1),
     datetime(2015, 6, 12),
 ]
-append_scm_df_base = ScmDataFrame(
+append_scm_df_base = OpenScmDataFrame(
     data=np.array([[1, 6.0, 7], [0.5, 3, 2], [2, 7, 0], [-1, -2, 3]]).T,
     index=append_scm_df_pairs_times,
     columns={
@@ -266,7 +266,7 @@ append_scm_df_base = ScmDataFrame(
 append_scm_df_pairs = [
     {
         "base": append_scm_df_base,
-        "other": ScmDataFrame(
+        "other": OpenScmDataFrame(
             data=np.array([[-1, 0, 1]]).T,
             index=append_scm_df_pairs_times,
             columns={
@@ -276,7 +276,7 @@ append_scm_df_pairs = [
             },
         ),
         "duplicate_rows": 1,
-        "expected": ScmDataFrame(
+        "expected": OpenScmDataFrame(
             data=np.array([[0, 3.0, 4], [0.5, 3, 2], [2, 7, 0], [-1, -2, 3]]).T,
             index=append_scm_df_pairs_times,
             columns={
@@ -288,7 +288,7 @@ append_scm_df_pairs = [
     },
     {
         "base": append_scm_df_base,
-        "other": ScmDataFrame(
+        "other": OpenScmDataFrame(
             data=np.array([[3, 3.5, 3.7], [1, 7, 11], [-2, 1, -1.4]]).T,
             index=append_scm_df_pairs_times,
             columns={
@@ -298,7 +298,7 @@ append_scm_df_pairs = [
             },
         ),
         "duplicate_rows": 1,
-        "expected": ScmDataFrame(
+        "expected": OpenScmDataFrame(
             data=np.array(
                 [
                     [2, 4.75, 5.35],
@@ -320,7 +320,7 @@ append_scm_df_pairs = [
     },
     {
         "base": append_scm_df_base,
-        "other": ScmDataFrame(
+        "other": OpenScmDataFrame(
             data=np.array([[3, 3.5, 3.7], [1, 7, 11], [-2, 1, -1.4], [-3, -4, -5]]).T,
             index=append_scm_df_pairs_times,
             columns={
@@ -335,7 +335,7 @@ append_scm_df_pairs = [
             },
         ),
         "duplicate_rows": 2,
-        "expected": ScmDataFrame(
+        "expected": OpenScmDataFrame(
             data=np.array(
                 [
                     [2, 4.75, 5.35],
@@ -357,7 +357,7 @@ append_scm_df_pairs = [
     },
     {
         "base": append_scm_df_base,
-        "other": ScmDataFrame(
+        "other": OpenScmDataFrame(
             data=np.array([[-1, 0, 1], [3, 4, 4.5], [0.1, 0.2, 0.3], [-4, -8, 10]]).T,
             index=append_scm_df_pairs_times,
             columns={
@@ -367,7 +367,7 @@ append_scm_df_pairs = [
             },
         ),
         "duplicate_rows": 4,
-        "expected": ScmDataFrame(
+        "expected": OpenScmDataFrame(
             data=np.array(
                 [[0, 3, 4], [1.75, 3.5, 3.25], [1.05, 3.6, 0.15], [-2.5, -5, 6.5]]
             ).T,
@@ -395,7 +395,7 @@ def iamdf_type():
 @pytest.fixture(scope="module")
 def rcp26():
     fname = join(TEST_DATA, "rcp26_emissions.csv")
-    return ScmDataFrame(fname)
+    return OpenScmDataFrame(fname)
 
 
 @pytest.fixture
@@ -575,7 +575,7 @@ def combo_df(request):
         assert len(vals["target"]) == len(combination.target_values)
         assert len(source) == len(combination.source_values)
 
-    df = ScmDataFrame(
+    df = OpenScmDataFrame(
         combination.source_values,
         columns={
             "scenario": ["a_scenario"],
